# openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: 镜中CMS
  version: 1.0.0
servers:
  - url: /
paths:
  /user/install:
    post:
      summary: 系统安装
      operationId: openapi_server.controllers.default_controller.install
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 安装成功
        '403':
          description: 已安装过
  /article/add:
    post:
      summary: 新增文章
      operationId: openapi_server.controllers.default_controller.article_add_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleContent'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  article_id:
                    type: integer

  /article/content:
    get:
      summary: 获取文章内容
      operationId: openapi_server.controllers.default_controller.article_content_get
      parameters:
        - name: article_id
          in: query
          description: 文章ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  article_id:
                    type: integer
                  name:
                    type: string
                  content:
                    type: string
                  category_id:
                    type: integer

  /article/del:
    post:
      summary: 删除文章
      operationId: openapi_server.controllers.default_controller.article_del_post
      parameters:
        - name: article_id
          in: query
          description: 文章ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /article/divide:
    post:
      summary: 划分文章到栏目
      operationId: openapi_server.controllers.default_controller.article_divide_post
      parameters:
        - name: article_id
          in: query
          description: 文章ID
          required: true
          schema:
            type: integer
        - name: category_id
          in: query
          description: 栏目ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /article/list:
    get:
      summary: 获取栏目内文章列表
      operationId: openapi_server.controllers.default_controller.article_list_get
      parameters:
        - name: category_id
          in: query
          description: 栏目ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'

  /category/add:
    post:
      summary: 新增栏目
      operationId: openapi_server.controllers.default_controller.category_add_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /category/del:
    post:
      summary: 删除栏目
      operationId: openapi_server.controllers.default_controller.category_del_post
      parameters:
        - name: category_id
          in: query
          description: 栏目ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /category/list:
    get:
      summary: 获取栏目列表
      operationId: openapi_server.controllers.default_controller.category_list_get
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

  /file/upload:
    post:
      summary: 文件上传
      operationId: openapi_server.controllers.default_controller.file_upload_post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string

  /file/download:
    post:
      summary: 文件下载
      operationId: openapi_server.controllers.default_controller.file_download_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  /system/add:
    post:
      summary: 新增系统字段
      operationId: openapi_server.controllers.default_controller.system_add_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setting'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /system/del:
    post:
      summary: 删除系统字段
      operationId: openapi_server.controllers.default_controller.system_del_post
      parameters:
        - name: key
          in: query
          description: 字段名
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  /system/list:
    get:
      summary: 获取系统字段
      operationId: openapi_server.controllers.default_controller.system_list_get
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Setting'

  /system/update:
    post:
      summary: 修改系统字段
      operationId: openapi_server.controllers.default_controller.system_update_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
              required:
                - key
                - value
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  updated:
                    type: boolean

# 移除 requestBodies 和不必要的 schemas
components:
  schemas:
    Category:
      type: object
      properties:
        category_id:
          type: integer
          description: ID 编号
        name:
          type: string
          description: 名称
      required:
        - category_id
        - name

    Article:
      type: object
      properties:
        article_id:
          type: integer
          description: ID 编号
        name:
          type: string
          description: 名称
        category_id:
          type: integer
          description: 分组
      required:
        - article_id
        - name
        - category_id

    ArticleContent:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
      required:
        - name
        - content

    Setting:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
      required:
        - key
        - value

# 移除不必要的组件定义